Caching of which blocks are supposed to be rendered, multi-threaded render was removed because it wasn't efficient with it, added area selection, selecting area, entity and saving the world now classified as Operations.

+++ Common/src/main/java/me/danetnaverno/editoni/util/Camera.java
@@ -2,5 +2,5 @@ package me.danetnaverno.editoni.util;
 
 public class Camera
 {
-    public static float x, y, z, pitch, yaw;
+    public static double x, y, z, pitch, yaw;
 }
